made the function a module and imported, instead of declaring inside the component
